backend fixed

+++ server/src/main/java/com/Cranco/Cranco/Notification/Notification.java
@@ -15,9 +15,10 @@ import java.util.Set;
 @Node("NOTIFICATION")
 public class Notification {
 
+//    notice :- only accepted ID types are string and long. Dont use same methods used in tutotials.
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     @Property("date")
     private Date date;
@@ -31,7 +32,7 @@ public class Notification {
     @Property("clicked")
     private Boolean clicked;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
     public Date getDate() {
@@ -47,7 +48,7 @@ public class Notification {
         return clicked;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
